document-processor/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2
python-multipart>=0.0.7
pytesseract>=0.3.10
pdfplumber>=0.11
pypdf>=3.9
Pillow>=10

